문의게시판 유저 글 수정, 글 삭제 유효성완료(지환)

+++ src/main/java/com/example/pj2be/controller/inquirycontroller/InquiryController.java
@@ -4,12 +4,17 @@ import com.example.pj2be.domain.answer.AnswerDTO;
 import com.example.pj2be.domain.inquiry.InquiryDTO;
 import com.example.pj2be.domain.member.MemberDTO;
 import com.example.pj2be.service.inquiryservice.InquiryService;
+import com.example.pj2be.utill.MemberAccess;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 import java.util.Map;
+
+import static com.example.pj2be.utill.MemberAccess.IsLoginMember;
+import static com.example.pj2be.utill.MemberAccess.MemberChecked;
 
 @RestController
 @RequiredArgsConstructor
@@ -29,8 +34,6 @@ public class InquiryController {
     @PostMapping("write")
     public ResponseEntity write(@RequestBody InquiryDTO dto) {
 
-//        로그인 객체 되면 수정
-
         if (!service.validate(dto)) {
             return ResponseEntity.badRequest().build();
         }
@@ -47,20 +50,49 @@ public class InquiryController {
     }
 
     @DeleteMapping("delete/{id}")
-    public ResponseEntity delete(@PathVariable Integer id) {
+    public ResponseEntity delete(@PathVariable Integer id,
+                                 @RequestBody InquiryDTO dto) {
 
-        service.delete(id);
-        return ResponseEntity.ok().build();
+        if (IsLoginMember(dto.getLogin_member_id())) {
+            System.out.println("로그인 되어있음: " + dto.getLogin_member_id());
+            if (MemberChecked(dto.getLogin_member_id(), dto.getInquiry_member_id()) == 0) {
+                System.out.println("아이디랑 멤버아이디 같음");
+                service.delete(id);
+                System.out.println("수정 완료");
+                return ResponseEntity.ok().build();
+            }
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+
+
+        if (MemberChecked(dto.getLogin_member_id(), dto.getInquiry_category()) == 2) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+
+        return ResponseEntity.internalServerError().build();
+
     }
 
     @PutMapping("edit")
     public ResponseEntity edit(@RequestBody InquiryDTO dto) {
 
-        System.out.println("문의 수정작동");
-        dto.setInquiry_member_id("testadmin");
+        if (IsLoginMember(dto.getLogin_member_id())) {
+            System.out.println("로그인 되어있음: " + dto.getLogin_member_id());
+            if (MemberChecked(dto.getLogin_member_id(), dto.getInquiry_member_id()) == 0) {
+                System.out.println("아이디랑 멤버아이디 같음");
+                service.update(dto);
+                System.out.println("수정 완료");
+                return ResponseEntity.ok().build();
+            }
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
 
-        service.update(dto);
-        return ResponseEntity.ok().build();
+
+        if (MemberChecked(dto.getLogin_member_id(), dto.getInquiry_category()) == 2) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+
+        return ResponseEntity.internalServerError().build();
     }
 
     @PostMapping("answer")
